Fix support for legacy terminals (Urxvt)
Thanks to u/katie-is-gay for tracking this down:
https://www.reddit.com/r/archlinux/comments/cnvlsa/lolcat_colors_glitched_only_at_urxvt/ewelhfn

And to @incomingstick for reporting it:
https://github.com/busyloop/lolcat/issues/99

Also thanks to @XVilka for maintaining an absurdly
comprehensive guide on Truecolor terminal support:
https://gist.github.com/XVilka/8346728

TLDR: We no longer rely on 24bit auto-detection by the Paint gem.
Lolcat now performs its own detection and otherwise defaults 256 colors.

The auto-detection is only a qualified guess.
Use the -t flag to ensure highest rainbow fidelity.

Fixes #97